move template builder to api.py
