+++ bluemira/codes/process/api.py
@@ -26,7 +26,7 @@ from dataclasses import dataclass
 from enum import Enum
 from importlib import resources
 from pathlib import Path
-from typing import Dict, List, TypeVar, Union
+from typing import Dict, List, Optional, TypeVar, Union
 
 from bluemira.base.look_and_feel import bluemira_print, bluemira_warn
 from bluemira.codes.error import CodesError
@@ -206,3 +206,129 @@ def _nested_check(process_name):
                 names += [_nested_check(p)]
             return list(flatten_iterable(names))
     return process_name
+
+
+class PROCESSTemplateBuilder:
+    """
+    An API patch to make PROCESS a little easier to work with before
+    the PROCESS team write a Python API.
+    """
+
+    def __init__(self):
+        self.values: Dict[str, float] = {}
+        self.bounds: Dict[str, Dict[str, str]] = {}
+        self.icc: List[int] = []
+        self.ixc: List[int] = []
+        self.minmax: int = 0
+        self.ioptimiz: int = 1
+
+    def set_minimisation_objective(self, name: str):
+        """
+        Set the minimisation objective equation to use when running PROCESS
+        """
+        minmax = OBJECTIVE_EQ_MAPPING.get(name, None)
+        if not minmax:
+            raise ValueError(f"There is no objective equation: '{name}'")
+
+        self.minmax = minmax
+
+    def set_maximisation_objective(self, name: str):
+        """
+        Set the maximisation objective equation to use when running PROCESS
+        """
+        minmax = OBJECTIVE_EQ_MAPPING.get(name, None)
+        if not minmax:
+            raise ValueError(f"There is no objective equation: '{name}'")
+        if minmax in OBJECTIVE_MIN_ONLY:
+            raise ValueError(
+                f"Equation {name} can only be used as a minimisation objective."
+            )
+        self.minmax = -minmax
+
+    def add_constraint(self, name: str):
+        """
+        Add a constraint to the PROCESS run
+        """
+        constraint = CONSTRAINT_EQ_MAPPING.get(name, None)
+        if not constraint:
+            raise ValueError(f"There is no constraint equation: '{name}'")
+        if constraint in self.icc:
+            bluemira_warn(f"Constraint {name} is already in the constraint list.")
+
+        if constraint in FV_CONSTRAINT_ITVAR_MAPPING.keys():
+            # Sensible (?) defaults. bounds are standard PROCESS for f-values for _most_
+            # f-value constraints.
+            self.add_fvalue_constraint(name, 0.5, 1e-3, 1.0)
+        else:
+            self.icc.append(constraint)
+
+    def add_fvalue_constraint(
+        self,
+        name: str,
+        value: float,
+        lower_bound: float = 1e-3,
+        upper_bound: float = 1.0,
+    ):
+        """
+        Add an f-value constraint to the PROCESS run
+        """
+        constraint = CONSTRAINT_EQ_MAPPING.get(name, None)
+        if not constraint:
+            raise ValueError(f"There is no constraint equation: '{name}'")
+
+        if constraint not in FV_CONSTRAINT_ITVAR_MAPPING.keys():
+            raise ValueError(f"Constraint '{name}' is not an f-value constraint.")
+
+        itvar = FV_CONSTRAINT_ITVAR_MAPPING[constraint]
+        if itvar not in self.ixc:
+            self.add_variable(
+                VAR_ITERATION_MAPPING[itvar], value, lower_bound, upper_bound
+            )
+
+    def add_variable(
+        self,
+        name: str,
+        value: float,
+        lower_bound: Optional[float] = None,
+        upper_bound: Optional[float] = None,
+    ):
+        """
+        Add an iteration variable to the PROCESS run
+        """
+        itvar = ITERATION_VAR_MAPPING.get(name, None)
+        if not itvar:
+            raise ValueError(f"There is no iteration variable: '{name}'")
+
+        if itvar in self.ixc:
+            bluemira_warn(
+                "Iterable variable {name} is already in the variable list. Updating value and bounds."
+            )
+            self.values[name] = value
+            if lower_bound:
+                self.bounds[str(itvar)]["l"] = lower_bound
+            if upper_bound:
+                self.bounds[str(itvar)]["u"] = upper_bound
+
+        else:
+            self.ixc.append(itvar)
+            self.values[name] = value
+
+        if lower_bound or upper_bound:
+            var_bounds = {}
+            if lower_bound:
+                var_bounds["l"] = lower_bound
+            if upper_bound:
+                var_bounds["u"] = upper_bound
+            self.bounds[str(itvar)] = var_bounds
+
+    def make_inputs(self) -> Dict[str, _INVariable]:
+        """
+        Make the ProcessInputs InVariable for the specified template
+        """
+        return ProcessInputs(
+            bounds=self.bounds,
+            icc=self.icc,
+            ixc=self.ixc,
+            minmax=self.minmax,
+            **self.values,
+        ).to_invariable()
